Replace seeds with seed references in all of the tags
I would like to be free to bump passwords and update private seeds
without having to worry about manually synchronizing the tags between
browsers on different devices.  I would like to take advantage of
Chrome sync to help me synchronize configuration between browsers.

The desire to use Chrome sync is at odds with the desire to keep my
secrets secret.  I don't want the secrets to be synchronized between
devices.  I'm okay doing that using some sort of sneaker net.  So, the
first step is to scrub the secrets from the individual tag records.

I needed to generate a key for each secret seed that could be used to
lookup the actual seed.  To do this, I generate a short digest of the
secret using an hmac.

The addition of options.salt is an added paranoid measure to ensure
that there is nothing secret about the reference.

Change-Id: I6a89fc8e6b75b8c2796c9cca41ea4c3a63b81a6d