Convert storage to use an asynchronous API
The motivation is to eventually use Chrome sync or a REST API to
synchronize urls and tags and all of the less-sensitive
configuration options.  Using either requires using an asynchronous
API.  It is possible to make a sychronous API behave like an
asynchronous API but not the other way around.  That is why I've
wrapped localStorage with an API that operations like an asynchronous
API using callbacks.

The effects of this change are far reaching.  I hesitated to make the
changes for some time until I realized that there was no way around
it.  In order to get the flexibility needed to use a different storage
mechanism to synchronize settings.

There are some errors in indentation on purpose.  The reason for this
is so that the diff from the previous version shows more in common
than not.  This can be fixed with a follow on commit that changes only
whitespace.

Change-Id: Iea34abc01d7c142afb75052e8138effe88b1dd04