When saving options, start with the previous set of options
The problem that I found was that I added the options for salt and
seeds which don't have a UI element.  I didn't feel a need to add a UI
element for either of these.  So, to preserve the values for salt and
seeds, I load the options before saving them.

Change-Id: Ief12f4e159f5fd19b4c2b0adfa792b25415f2a43