Don't do anything with pages that don't have password fields
I saw this problem on Discover's page.  In debugging, I found that
there were 6 instances of content-script.js each with a different url.
Only one of them was the page with the password field.  The problem
was when I changed things in the extension popup, it would send a
message to the wrong content instance.  This tweak seems to fix it.

Change-Id: I1eae2ee1d7cfbe7c004d37a17619e7edaabd05f0