Remove the version from the DB
I was trying to figure out if the version of the DB should be stored
locally or remotely.  It occurred to me that since part of the DB is
shared and part is local to a particular instance of the browser that
the notion of a single version of the DB doesn't make sense.  So, I
need a strategy for handling versioning that is more flexible.

So, actually, I don't store the version at all anymore.  Instead, the code
figures out if anything in the data needs to be updated to fit the current
structure by looking at the data itself.  We'll need to be sure that upgrades
happen fairly uniformly across browsers or old versions may have a problem with
the new data.

A version stored on the remote might still be useful for this but
we'll see what happens.

I also abandoned all the migrations up to version 5 of the DB.  If
you're coming from password hasher plus, you'll need to get up to
version 5 and then bring it over.  If password hasher plus goes beyond
version 5 then it won't be possible to bring your DB over.

Change-Id: Ie167fa30bd6bbfa099f94e65a2bcd0b204127bd1