Combine tags with url
Honestly, I don't think that url and tag was the right way to split
this data.  To me, it is more natural to attach the metadata to the
url and not the tag.  I don't think it is right for every url using
the same tag to use the same length, strength, and seed.

I've wondered about this for a while but the real motivation behind
combining them is that chrome sync only allows 512 records and taking
one for url and another for tag is wasteful.

Change-Id: Iee1d8125e029b89724980d171d61eaf2c8655ba3