Get the url from the port rather than the message
I want to trim down the amount of data exchange between the
content-script and the background page because I tend to distrust the
content-script.  So, I noticed that the url can be obtained from the
port which I tend to trust more before it is the extension plumbing
that adds it.

Change-Id: Ide2447d4907227f519a8bd1b8bf74375113e13a2