Don't send the secrets to the content script
I think this was a security hole to let the secrets get passed over
to the content script.  I didn't try but I had a feeling that a
malicious script in a login page could grab the secrets and make off
with them.  I feel much better now that the background script will not
send secrets over to the content script.

Change-Id: I276247435cc2236108b7c695d167c79038c2b394